requirements.txt
pypdf==4.2.0
pillow==10.3.0